Remove unnecessary plumbing in PluginManager for commands
