+++ browser/src/Services/CommandManager.ts
@@ -10,20 +10,7 @@ import * as Log from "./../Log"
 import { INeovimInstance } from "./../neovim"
 import { ITask, ITaskProvider } from "./Tasks"
 
-export type ICommandCallback = (args?: any) => any
-export type ICommandEnabledCallback = () => boolean
-
-export interface ICommand {
-    command: string
-    name: string
-    detail: string
-    enabled?: ICommandEnabledCallback
-    messageSuccess?: string
-    messageFail?: string
-    execute: ICommandCallback
-}
-
-export class CallbackCommand implements ICommand {
+export class CallbackCommand implements Oni.ICommand {
     public messageSuccess?: string
     public messageFail?: string
 
@@ -31,12 +18,12 @@ export class CallbackCommand implements ICommand {
         public command: string,
         public name: string,
         public detail: string,
-        public execute: ICommandCallback,
-        public enabled?: ICommandEnabledCallback) {
+        public execute: Oni.ICommandCallback,
+        public enabled?: Oni.ICommandEnabledCallback) {
     }
 }
 
-export class VimCommand implements ICommand {
+export class VimCommand implements Oni.ICommand {
     constructor(
         public command: string,
         public name: string, public detail: string,
@@ -52,13 +39,13 @@ export class VimCommand implements ICommand {
 
 export class CommandManager implements ITaskProvider {
 
-    private _commandDictionary: { [key: string]: ICommand } = {}
+    private _commandDictionary: { [key: string]: Oni.ICommand } = {}
 
     public clearCommands(): void {
         this._commandDictionary = {}
     }
 
-    public registerCommand(command: ICommand): void {
+    public registerCommand(command: Oni.ICommand): void {
         if (this._commandDictionary[command.command]) {
             Log.error(`Tried to register multiple commands for: ${command.name}`)
             return
