+++ browser/src/Services/Commands.ts
@@ -24,7 +24,7 @@ import { windowManager } from "./../Services/WindowManager"
 
 import * as UI from "./../UI/index"
 
-import { CallbackCommand, CommandManager, ICommandCallback } from "./CommandManager"
+import { CallbackCommand, CommandManager } from "./CommandManager"
 
 import * as Platform from "./../Platform"
 import { replaceAll } from "./../Utility"
@@ -108,7 +108,7 @@ export const registerBuiltInCommands = (commandManager: CommandManager, pluginMa
  * Higher-order function for commands dealing with completion
  * - checks that the completion menu is open
  */
-const autoCompletionCommand = (innerCommand: ICommandCallback) => {
+const autoCompletionCommand = (innerCommand: Oni.ICommandCallback) => {
     return () => {
         if (UI.Selectors.areCompletionsVisible()) {
             return innerCommand()
@@ -126,7 +126,7 @@ const previousCompletionItem = autoCompletionCommand(() => {
     UI.Actions.previousCompletion()
 })
 
-const popupMenuCommand = (innerCommand: ICommandCallback) => {
+const popupMenuCommand = (innerCommand: Oni.ICommandCallback) => {
     return () => {
         if (UI.Selectors.isPopupMenuOpen()) {
             return innerCommand()
